solvers : Separating General Functionalities(polynomial solver) into `solvers` Submodule (#2465)
* adding polynomial solver

* fix : handle zero discriminant case

* test cases for quadratic solver

* adding tests for cubic solver

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Removed redundant cases, auto-handled by initialized solutions.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------
