+++ kornia/geometry/solvers/__init__.py
@@ -1,0 +1,3 @@
+from .polynomial_solver import solve_cubic, solve_quadratic
+
+__all__ = ["solve_quadratic", "solve_cubic"]
